web-animations/animatable/animatable-animate-timing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animatable.animate() method timing argument test</title>
<meta name="assert" content="The timing parameters of the animation. This value is passed to the Animation constructor as the timing parameter and has the same interpretation as defined for that constructor">
<link rel="help" href="https://web-animations.github.io/web-animations-spec/bikeshed/Overview.html#the-animatable-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animatable = createDiv(this);
    var player = animatable.animate(null);

    assert_timing_equals(player.source.timing, DEFAULT_TIMING);
}, 'By default Animatable.animate() call produces anumation with default timing properties');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var animatable = createDiv(this);
    var player = animatable.animate(null, timing);

    assert_timing_equals(player.source.timing, timing);
}, 'Animatable.animate() call with specified timing argument produces anumation ' +
    'with specified timing properties');

test(function() {
    var animatable = createDiv(this);
    var player = animatable.animate(null, 5);

    var timing = Object.create(DEFAULT_TIMING);
    timing.duration = 5;
    assert_timing_equals(player.source.timing, timing);
}, 'Animatable.animate() call with double timing argument ' +
    'creates Animation instance with default timing properties and specified duration');
</script>
</body>
